+++ src/main/java/com/brewery/application/dto/outputdto/RatingOutDto.java
@@ -16,7 +16,7 @@ public class RatingOutDto {
 
     private String comment;
 
-    private Double rating;
+    private Float rating;
 
     private String itemName;
