+++ src/main/java/com/brewery/application/entity/Invoice.java
@@ -17,12 +17,12 @@ public class Invoice {
     @GeneratedValue(strategy = GenerationType.UUID)
     private UUID id;
 
-    private Double amount;
+    private Float amount;
 
-    private Double deliveryFee;
+    private Float deliveryFee;
 
-    private Double gst;
+    private Float gst;
 
-    private Double totalAmount;
+    private Float totalAmount;
 
 }
